requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aniso8601==1.3.0          # via flask-restful
arrow==0.10.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7                # via flask
flask-restful==0.3.6
flask-sqlalchemy==2.3.0
flask==0.12.2
idna==2.6                 # via requests
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
markupsafe==1.0           # via jinja2
psycopg2==2.7.3.1
python-crontab==2.2.4
python-dateutil==2.6.1    # via aniso8601, arrow, python-crontab
pytz==2017.2              # via flask-restful
requests==2.18.4
six==1.11.0               # via flask-restful, python-dateutil, sqlalchemy-utils
sqlalchemy-utils==0.32.17
sqlalchemy==1.1.14
urllib3==1.22             # via requests
uwsgi==2.0.15
werkzeug==0.12.2          # via flask
